Trace raw SQL statements.
This is done using the ExecutionContext as storage place
(since it is an unique object shared for a single instance).
It is an ideal place as raw statement simply come as a string,
so no way to store the current span there.

We still try to ignore PRAGMA statements, though.